// src/components/table/enum.ts
/*
 * @Description: Copyright (c) ydfk. All rights reserved
 * @Author: ydfk
 * @Date: 2022-10-27 15:14:58
 * @LastEditors: zhaozijun zaozijun@foxmail.com
 * @LastEditTime: 2023-11-27 16:28:39
 */
export enum TableActionKeyEnum {
  EDIT = "edit",
  DETAIL = "detail",
  DELETE = "delete",
  DISABLE = "disable",
  ENABLE = "enable",
  REMOVE = "remove",
  RESET_PASSWORD = "resetPassword",
  DOWNLOAD = "download",
  SELECT = "select",
  SELECTED = "selected",
}